BUGFIX: Resolve recursive dependency
